(refactor) split MCP server into per-domain tool modules (#168)
Split the monolithic server.ts (1198 lines) into 7 focused domain
modules under tools/, each with co-located tests. server.ts is now a
thin orchestrator that imports and calls register functions.
