+++ packages/mcp/src/tools/posts.ts
@@ -1,0 +1,415 @@
+// SPDX-License-Identifier: AGPL-3.0-only
+// Copyright (C) 2026 Oleksii PELYKH
+
+import type { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
+import { z } from "zod";
+import { readFile, stat } from "node:fs/promises";
+import { extname } from "node:path";
+
+import {
+  resolveConfig,
+  LinkedInClient,
+  getCurrentPersonUrn,
+  createPost,
+  getPost,
+  listPosts,
+  updatePost,
+  deletePost,
+  uploadImage,
+  uploadVideo,
+  uploadDocument,
+  listOrganizations,
+  SUPPORTED_IMAGE_TYPES,
+  DOCUMENT_EXTENSIONS,
+  DOCUMENT_MAX_SIZE_BYTES,
+} from "@linkedctl/core";
+import type { PostContent, PostLifecycleState } from "@linkedctl/core";
+
+export function registerPostTools(server: McpServer): void {
+  server.registerTool(
+    "post_create",
+    {
+      title: "Create Post",
+      description:
+        "Create a post on LinkedIn with optional media attachment. Supports URNs for pre-uploaded media or local file paths for combined upload+post.",
+      inputSchema: {
+        text: z.string().describe("The text content of the post"),
+        visibility: z.enum(["PUBLIC", "CONNECTIONS"]).optional().describe("Post visibility (defaults to PUBLIC)"),
+        draft: z.boolean().optional().describe("Save as draft instead of publishing (defaults to false)"),
+        image: z.string().optional().describe("Image URN to attach (e.g. urn:li:image:C5608AQ...)"),
+        video: z.string().optional().describe("Video URN to attach (e.g. urn:li:video:D5608AQ...)"),
+        document: z.string().optional().describe("Document URN to attach (e.g. urn:li:document:D123...)"),
+        article_url: z.string().optional().describe("Article URL to attach"),
+        images: z.array(z.string()).optional().describe("Array of image URNs for multi-image post (minimum 2)"),
+        image_file: z.string().optional().describe("Path to a local image file to upload and attach"),
+        video_file: z.string().optional().describe("Path to a local video file to upload and attach"),
+        document_file: z.string().optional().describe("Path to a local document file to upload and attach"),
+        image_files: z
+          .array(z.string())
+          .optional()
+          .describe("Array of local image file paths to upload and attach (minimum 2)"),
+        poll: z.string().optional().describe("Poll question text (creates a poll post)"),
+        poll_options: z
+          .array(z.string())
+          .optional()
+          .describe("Array of poll answer options (minimum 2, maximum 4; required when poll is set)"),
+        poll_duration: z
+          .enum(["ONE_DAY", "THREE_DAYS", "ONE_WEEK", "TWO_WEEKS"])
+          .optional()
+          .describe("How long the poll stays open (defaults to THREE_DAYS)"),
+        as_org: z
+          .string()
+          .optional()
+          .describe("Post as an organization (numeric organization ID). The authenticated user must be an admin."),
+        profile: z.string().optional().describe("Profile name to use from config file"),
+      },
+    },
+    async (args) => {
+      const contentFlags = [
+        args.image,
+        args.video,
+        args.document,
+        args.article_url,
+        args.images,
+        args.image_file,
+        args.video_file,
+        args.document_file,
+        args.image_files,
+        args.poll,
+      ].filter((v) => v !== undefined);
+      if (contentFlags.length > 1) {
+        return {
+          content: [
+            {
+              type: "text" as const,
+              text: "Only one content option may be specified: image, video, document, article_url, images, image_file, video_file, document_file, image_files, or poll",
+            },
+          ],
+          isError: true,
+        };
+      }
+
+      let postContent: PostContent | undefined;
+      if (args.image !== undefined) {
+        postContent = { media: { id: args.image } };
+      } else if (args.video !== undefined) {
+        postContent = { media: { id: args.video } };
+      } else if (args.document !== undefined) {
+        postContent = { media: { id: args.document } };
+      } else if (args.article_url !== undefined) {
+        postContent = { article: { source: args.article_url } };
+      } else if (args.images !== undefined) {
+        if (args.images.length < 2) {
+          return {
+            content: [{ type: "text" as const, text: "Multi-image posts require at least 2 image URNs" }],
+            isError: true,
+          };
+        }
+        postContent = { multiImage: { images: args.images.map((id) => ({ id })) } };
+      } else if (args.poll !== undefined) {
+        const pollOptions = args.poll_options ?? [];
+        if (pollOptions.length < 2) {
+          return {
+            content: [{ type: "text" as const, text: "Poll posts require at least 2 poll_options" }],
+            isError: true,
+          };
+        }
+        if (pollOptions.length > 4) {
+          return {
+            content: [{ type: "text" as const, text: "Poll posts allow at most 4 poll_options" }],
+            isError: true,
+          };
+        }
+        postContent = {
+          poll: {
+            question: args.poll,
+            options: pollOptions.map((text) => ({ text })),
+            settings: { duration: args.poll_duration ?? "THREE_DAYS" },
+          },
+        };
+      }
+
+      const requiredScopes = ["openid", "profile", "email", "w_member_social"];
+      if (args.as_org !== undefined) {
+        requiredScopes.push("w_organization_social");
+      }
+
+      const { config } = await resolveConfig({
+        profile: args.profile,
+        requiredScopes,
+      });
+      // resolveConfig guarantees oauth.accessToken and apiVersion are defined
+      const accessToken = config.oauth?.accessToken ?? "";
+      const apiVersion = config.apiVersion ?? "";
+      const client = new LinkedInClient({ accessToken, apiVersion });
+
+      let authorUrn: string;
+      if (args.as_org !== undefined) {
+        const orgUrn = `urn:li:organization:${args.as_org}`;
+        const response = await listOrganizations(client, { count: 100 });
+        const isAdmin = response.elements.some((acl) => acl.organization === orgUrn);
+        if (!isAdmin) {
+          return {
+            content: [
+              {
+                type: "text" as const,
+                text: `You are not an administrator of organization ${args.as_org}`,
+              },
+            ],
+            isError: true,
+          };
+        }
+        authorUrn = orgUrn;
+      } else {
+        authorUrn = await getCurrentPersonUrn(client);
+      }
+
+      // Handle file-based uploads if no URN-based content was resolved
+      if (postContent === undefined) {
+        if (args.image_file !== undefined) {
+          const ext = extname(args.image_file).toLowerCase();
+          const contentType = SUPPORTED_IMAGE_TYPES.get(ext);
+          if (contentType === undefined) {
+            const supported = [...SUPPORTED_IMAGE_TYPES.keys()].join(", ");
+            return {
+              content: [
+                {
+                  type: "text" as const,
+                  text: `Unsupported image format "${ext}". Supported formats: ${supported}`,
+                },
+              ],
+              isError: true,
+            };
+          }
+          const data = new Uint8Array(await readFile(args.image_file));
+          const urn = await uploadImage(client, { owner: authorUrn, data, contentType });
+          postContent = { media: { id: urn } };
+        } else if (args.video_file !== undefined) {
+          const fileStat = await stat(args.video_file);
+          if (!fileStat.isFile()) {
+            return {
+              content: [{ type: "text" as const, text: `Not a file: ${args.video_file}` }],
+              isError: true,
+            };
+          }
+          const data = await readFile(args.video_file);
+          const urn = await uploadVideo(client, { owner: authorUrn, data });
+          postContent = { media: { id: urn } };
+        } else if (args.document_file !== undefined) {
+          const ext = extname(args.document_file).toLowerCase();
+          if (!DOCUMENT_EXTENSIONS.includes(ext as (typeof DOCUMENT_EXTENSIONS)[number])) {
+            return {
+              content: [
+                {
+                  type: "text" as const,
+                  text: `Unsupported file type "${ext}". Supported types: ${DOCUMENT_EXTENSIONS.join(", ")}`,
+                },
+              ],
+              isError: true,
+            };
+          }
+          const fileStat = await stat(args.document_file);
+          if (fileStat.size > DOCUMENT_MAX_SIZE_BYTES) {
+            const sizeMB = Math.round(fileStat.size / (1024 * 1024));
+            return {
+              content: [{ type: "text" as const, text: `File is ${sizeMB} MB, which exceeds the 100 MB limit.` }],
+              isError: true,
+            };
+          }
+          const data = new Uint8Array(await readFile(args.document_file));
+          const urn = await uploadDocument(client, { owner: authorUrn, data });
+          postContent = { media: { id: urn } };
+        } else if (args.image_files !== undefined) {
+          if (args.image_files.length < 2) {
+            return {
+              content: [{ type: "text" as const, text: "Multi-image file posts require at least 2 image file paths" }],
+              isError: true,
+            };
+          }
+          const urns: string[] = [];
+          for (const filePath of args.image_files) {
+            const ext = extname(filePath).toLowerCase();
+            const contentType = SUPPORTED_IMAGE_TYPES.get(ext);
+            if (contentType === undefined) {
+              const supported = [...SUPPORTED_IMAGE_TYPES.keys()].join(", ");
+              return {
+                content: [
+                  {
+                    type: "text" as const,
+                    text: `Unsupported image format "${ext}" for file "${filePath}". Supported formats: ${supported}`,
+                  },
+                ],
+                isError: true,
+              };
+            }
+            const data = new Uint8Array(await readFile(filePath));
+            const urn = await uploadImage(client, { owner: authorUrn, data, contentType });
+            urns.push(urn);
+          }
+          postContent = { multiImage: { images: urns.map((id) => ({ id })) } };
+        }
+      }
+
+      const visibility = args.visibility ?? "PUBLIC";
+      const lifecycleState: PostLifecycleState = args.draft === true ? "DRAFT" : "PUBLISHED";
+
+      const postUrn = await createPost(client, {
+        author: authorUrn,
+        text: args.text,
+        visibility,
+        content: postContent,
+        lifecycleState,
+      });
+
+      return {
+        content: [{ type: "text" as const, text: `Post created: ${postUrn}` }],
+      };
+    },
+  );
+
+  server.registerTool(
+    "post_get",
+    {
+      title: "Get Post",
+      description: "Fetch a single LinkedIn post by URN",
+      inputSchema: {
+        urn: z.string().describe("Post URN (e.g. urn:li:share:123)"),
+        profile: z.string().optional().describe("Profile name to use from config file"),
+      },
+    },
+    async (args) => {
+      const { config } = await resolveConfig({
+        profile: args.profile,
+        requiredScopes: ["openid", "profile", "email", "w_member_social"],
+      });
+      const accessToken = config.oauth?.accessToken ?? "";
+      const apiVersion = config.apiVersion ?? "";
+      const client = new LinkedInClient({ accessToken, apiVersion });
+
+      const post = await getPost(client, args.urn);
+
+      return {
+        content: [{ type: "text" as const, text: JSON.stringify(post, null, 2) }],
+      };
+    },
+  );
+
+  server.registerTool(
+    "post_list",
+    {
+      title: "List Posts",
+      description: "List LinkedIn posts with pagination",
+      inputSchema: {
+        count: z.number().optional().describe("Number of posts to return (default 10, max 100)"),
+        start: z.number().optional().describe("Starting index for pagination (default 0)"),
+        as_org: z
+          .string()
+          .optional()
+          .describe(
+            "List posts of an organization (numeric organization ID). The authenticated user must be an admin.",
+          ),
+        profile: z.string().optional().describe("Profile name to use from config file"),
+      },
+    },
+    async (args) => {
+      const requiredScopes = ["openid", "profile", "email", "w_member_social"];
+      if (args.as_org !== undefined) {
+        requiredScopes.push("r_organization_social");
+      }
+
+      const { config } = await resolveConfig({
+        profile: args.profile,
+        requiredScopes,
+      });
+      const accessToken = config.oauth?.accessToken ?? "";
+      const apiVersion = config.apiVersion ?? "";
+      const client = new LinkedInClient({ accessToken, apiVersion });
+
+      let authorUrn: string;
+      if (args.as_org !== undefined) {
+        const orgUrn = `urn:li:organization:${args.as_org}`;
+        const orgResponse = await listOrganizations(client, { count: 100 });
+        const isAdmin = orgResponse.elements.some((acl) => acl.organization === orgUrn);
+        if (!isAdmin) {
+          return {
+            content: [
+              {
+                type: "text" as const,
+                text: `You are not an administrator of organization ${args.as_org}`,
+              },
+            ],
+            isError: true,
+          };
+        }
+        authorUrn = orgUrn;
+      } else {
+        authorUrn = await getCurrentPersonUrn(client);
+      }
+
+      const response = await listPosts(client, {
+        author: authorUrn,
+        count: args.count,
+        start: args.start,
+      });
+
+      return {
+        content: [{ type: "text" as const, text: JSON.stringify(response, null, 2) }],
+      };
+    },
+  );
+
+  server.registerTool(
+    "post_update",
+    {
+      title: "Update Post",
+      description: "Update the commentary text of an existing LinkedIn post",
+      inputSchema: {
+        urn: z.string().describe("Post URN (e.g. urn:li:share:123)"),
+        text: z.string().describe("New text content for the post"),
+        profile: z.string().optional().describe("Profile name to use from config file"),
+      },
+    },
+    async (args) => {
+      const { config } = await resolveConfig({
+        profile: args.profile,
+        requiredScopes: ["openid", "profile", "email", "w_member_social"],
+      });
+      const accessToken = config.oauth?.accessToken ?? "";
+      const apiVersion = config.apiVersion ?? "";
+      const client = new LinkedInClient({ accessToken, apiVersion });
+
+      await updatePost(client, args.urn, { text: args.text });
+
+      return {
+        content: [{ type: "text" as const, text: `Post updated: ${args.urn}` }],
+      };
+    },
+  );
+
+  server.registerTool(
+    "post_delete",
+    {
+      title: "Delete Post",
+      description: "Delete a LinkedIn post by URN",
+      inputSchema: {
+        urn: z.string().describe("Post URN (e.g. urn:li:share:123)"),
+        profile: z.string().optional().describe("Profile name to use from config file"),
+      },
+    },
+    async (args) => {
+      const { config } = await resolveConfig({
+        profile: args.profile,
+        requiredScopes: ["openid", "profile", "email", "w_member_social"],
+      });
+      const accessToken = config.oauth?.accessToken ?? "";
+      const apiVersion = config.apiVersion ?? "";
+      const client = new LinkedInClient({ accessToken, apiVersion });
+
+      await deletePost(client, args.urn);
+
+      return {
+        content: [{ type: "text" as const, text: `Post deleted: ${args.urn}` }],
+      };
+    },
+  );
+}
