+++ packages/mcp/src/tools/test-setup.ts
@@ -1,0 +1,44 @@
+// SPDX-License-Identifier: AGPL-3.0-only
+// Copyright (C) 2026 Oleksii PELYKH
+
+import { beforeEach, afterEach, vi } from "vitest";
+import { Client } from "@modelcontextprotocol/sdk/client/index.js";
+import { InMemoryTransport } from "@modelcontextprotocol/sdk/inMemory.js";
+import { createMcpServer } from "../server.js";
+
+/**
+ * Shared test context for MCP tool tests.
+ * Call `setupMcpTestClient()` inside your describe block to get a `client` accessor.
+ *
+ * IMPORTANT: Each test file must declare its own `vi.mock(...)` blocks at the top level
+ * because vitest hoists them per-file.
+ */
+export function setupMcpTestClient(): { getClient: () => Client } {
+  let client: Client;
+  let cleanup: () => Promise<void>;
+
+  beforeEach(async () => {
+    vi.clearAllMocks();
+
+    const server = createMcpServer();
+    const [clientTransport, serverTransport] = InMemoryTransport.createLinkedPair();
+
+    client = new Client({ name: "test-client", version: "1.0.0" });
+
+    await server.connect(serverTransport);
+    await client.connect(clientTransport);
+
+    cleanup = async () => {
+      await client.close();
+      await server.close();
+    };
+  });
+
+  afterEach(async () => {
+    await cleanup();
+  });
+
+  return {
+    getClient: () => client,
+  };
+}
